+++ src/app/JWT4B.java
@@ -2,7 +2,6 @@ package app;
 
 public class JWT4B {
 	public static void main(String[] args){
-		// TODO replace deprecated observers
 		// This solves the export warning concerning not finding a main method	
 		System.out.println("This JAR is meant to be run through BURP extender");
 	}
