+++ src/app/algorithm/AlgorithmType.java
@@ -1,6 +1,7 @@
 package app.algorithm;
 
 public class AlgorithmType {
+	// dummy class that holds the different algorithm "types"
 	public final static String none = "none";
 	public final static String symmetric = "symmetric";
 	public final static String asymmetric = "asymmetric";
