+++ src/app/helpers/KeyValuePair.java
@@ -24,5 +24,4 @@ public class KeyValuePair {
 	public void setValue(String value) {
 		this.value = value;
 	}
-
 }
