+++ src/app/tokenposition/AuthorizationBearerHeader.java
@@ -2,11 +2,10 @@ package app.tokenposition;
 
 import java.util.List;
 
-import javax.swing.JOptionPane;
-
 import app.helpers.Config;
-import app.helpers.CustomJWToken;
+import model.CustomJWToken;
 
+// finds and replaces JWT's in authorization headers
 public class AuthorizationBearerHeader extends ITokenPosition {
 
 	private String selectedKeyword;
