+++ src/model/CustomJWToken.java
@@ -1,4 +1,4 @@
-package app.helpers;
+package model;
 
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
@@ -7,8 +7,6 @@ import java.util.Date;
 import java.util.List;
 import java.util.Map;
 import java.util.TimeZone;
-
-import model.TimeClaim;
 
 import org.apache.commons.codec.binary.Base64;
 import org.apache.commons.codec.binary.StringUtils;
@@ -23,6 +21,8 @@ import com.eclipsesource.json.JsonValue;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
+
+import app.helpers.Output;
 
 /* 
  * This Class is implemented separately to get raw access to the content of the Tokens. 
@@ -187,7 +187,6 @@ public class CustomJWToken extends JWT {
 	@Override
 	public String getToken() {
 		String content = String.format("%s.%s", b64(jsonMinify(getHeaderJson())), b64(jsonMinify((getPayloadJson()))));
-
 		String signatureEncoded = Base64.encodeBase64URLSafeString(this.signature);
 
 		return String.format("%s.%s", content, signatureEncoded);
@@ -219,8 +218,7 @@ public class CustomJWToken extends JWT {
 			parts = new String[] { parts[0], parts[1], "" };
 		}
 		if (parts.length != 3) {
-			throw new JWTDecodeException(
-					String.format("The token was expected to have 3 parts, but got %s.", parts.length));
+			throw new JWTDecodeException(String.format("The token was expected to have 3 parts, but got %s.", parts.length));
 		}
 		return parts;
 	}
@@ -283,8 +281,7 @@ public class CustomJWToken extends JWT {
 		String algorithm = "";
 		try {
 			algorithm = getHeaderJsonNode().get("alg").asText();
-		} catch (Exception e) {
-		}
+		} catch (Exception e) { }
 		return algorithm;
 	}
