+++ src/model/Strings.java
@@ -1,4 +1,4 @@
-package app.helpers;
+package model;
 
 import java.io.BufferedReader;
 import java.io.FileReader;
